Refactored methods that were crowding the activity
