+++ app/src/main/java/com/litlgroup/litl/utils/Permissions.java
@@ -1,10 +1,60 @@
 package com.litlgroup.litl.utils;
 
+import android.app.Activity;
+import android.content.pm.PackageManager;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
+import android.widget.Toast;
+
+import com.litlgroup.litl.Manifest;
+
 /**
  * Created by Hari on 8/26/2016.
  */
 public class Permissions {
 
+
+    Activity activity;
+
+    public Permissions(Activity activity)
+    {
+        this.activity = activity;
+    }
+
+    public boolean checkPermissionForCamera(){
+        int result = ContextCompat.checkSelfPermission(activity, Manifest.permission.CAMERA);
+        if (result == PackageManager.PERMISSION_GRANTED){
+            return true;
+        } else {
+            return false;
+        }
+
+    } public static final int CAMERA_PERMISSION_REQUEST_CODE = 3;
+    public void requestPermissionForCamera(){
+        if (ActivityCompat.shouldShowRequestPermissionRationale(activity, Manifest.permission.CAMERA)){
+            Toast.makeText(activity, "Camera permission is required, please allow in App Settings", Toast.LENGTH_LONG).show();
+        } else {
+            ActivityCompat.requestPermissions(activity, new String[]{Manifest.permission.CAMERA},CAMERA_PERMISSION_REQUEST_CODE);
+        }
+    }
+
+
+    public boolean checkPermissionForExternalStorage(){
+        int result = ContextCompat.checkSelfPermission(activity, Manifest.permission.WRITE_EXTERNAL_STORAGE);
+        if (result == PackageManager.PERMISSION_GRANTED){
+            return true;
+        } else {
+            return false;
+        }
+    }
+    public static final int EXTERNAL_STORAGE_PERMISSION_REQUEST_CODE = 2;
+    public void requestPermissionForExternalStorage(){
+        if (ActivityCompat.shouldShowRequestPermissionRationale(activity, Manifest.permission.WRITE_EXTERNAL_STORAGE)){
+            Toast.makeText(activity, "External Storage permission is required. Please allow in App Settings", Toast.LENGTH_LONG).show();
+        } else {
+            ActivityCompat.requestPermissions(activity, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},EXTERNAL_STORAGE_PERMISSION_REQUEST_CODE);
+        }
+    }
 
 
 }
